Refactor constant_iterator
